Simple Editing modal with society data piped in
I thiiiink I'm doing this right? The dialog looks ugly but I believe I did the scaffolding right in the html.js (line 237)  and by making a corresponding DialogEditSociety component. Does that seem correct?

Not sure how to do the nice trick where it loads over the background like the create dialog does, but that's for later.